requirements.txt
django>=1.3
djangorestframework>=2.3
